+++ bridge/client/app/_interfaces/metadata.ts
