+++ bridge/client/app/_services/api.service.mock.ts
@@ -14,7 +14,7 @@ import { WebhookConfig, WebhookConfigMethod } from '../../../shared/interfaces/w
 import { UniformRegistrationLogResponse } from '../../../shared/interfaces/uniform-registration-log';
 import { Secret } from '../_models/secret';
 import { SecretScope } from '../../../shared/interfaces/secret-scope';
-import { IMetadata } from '../_models/IMetadata';
+import { IMetadata } from '../_interfaces/metadata';
 import { FileTree } from '../../../shared/interfaces/resourceFileTree';
 import { HttpResponse } from '@angular/common/http';
 import { SequenceResult } from '../_models/sequence-result';
