+++ bridge/client/app/_services/data.service.ts
@@ -33,7 +33,7 @@ import { SecretScope } from '../../../shared/interfaces/secret-scope';
 import { IGitDataExtended } from '../_interfaces/git-upstream';
 import { getGitData } from '../_utils/git-upstream.utils';
 import { ICustomSequences } from '../../../shared/interfaces/custom-sequences';
-import { IMetadata } from '../_models/IMetadata';
+import { IMetadata } from '../_interfaces/metadata';
 
 @Injectable({
   providedIn: 'root',
