+++ bridge/client/app/app-header/app-header.component.ts
@@ -14,7 +14,7 @@ import { KeptnInfo } from '../_models/keptn-info';
 import { DtSwitchChange } from '@dynatrace/barista-components/switch';
 import { VersionInfo } from '../_models/keptn-versions';
 import { DtSelect } from '@dynatrace/barista-components/select';
-import { IMetadata } from '../_models/IMetadata';
+import { IMetadata } from '../_interfaces/metadata';
 
 @Component({
   selector: 'ktb-header',
